[Authetication] Removed auth from porject
