+++ src/LinuxLudo.API/Controllers/GamesController.cs
@@ -5,12 +5,10 @@ using LinuxLudo.API.Domain.Models;
 using LinuxLudo.API.Domain.Resources;
 using LinuxLudo.API.Domain.Response;
 using LinuxLudo.API.Domain.Services;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LinuxLudo.API.Controllers
 {
-    [Authorize]
     [Route("/api/[controller]")]
     public class GamesController : Controller
     {
