+++ src/LinuxLudo.API/Database/Context/AppDbContext.cs
@@ -2,17 +2,14 @@ using System;
 using LinuxLudo.API.Database.Configurations;
 using LinuxLudo.API.Domain.Enums;
 using LinuxLudo.API.Domain.Models;
-using LinuxLudo.API.Domain.Models.Auth;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
 namespace LinuxLudo.API.Database.Context
 {
-    public class AppDbContext : IdentityDbContext<User, Role, Guid>
+    public class AppDbContext : DbContext
     {
         public DbSet<Game> Games { get; set; }
-        public DbSet<GamePlayerPivot> GamePlayerPivot { get; set; }
-        public DbSet<PlayerStats> PlayerStats { get; set; }
+        public DbSet<GameResult> Leaderboard { get; set; }
 
         public AppDbContext(DbContextOptions<AppDbContext> opts) : base(opts)
         { }
@@ -23,8 +20,7 @@ namespace LinuxLudo.API.Database.Context
             builder.HasPostgresExtension("uuid-ossp");
             builder.HasPostgresEnum<GameStates>();
             builder.ApplyConfiguration(new GameConfiguration());
-            builder.ApplyConfiguration(new PlayerStatsConfiguration());
-            builder.ApplyConfiguration(new GamePlayerPivotConfiguration());
+            builder.ApplyConfiguration(new GameResultConfiguration());
         }
     }
 }
