--- src/LinuxLudo.API/Domain/Models/Auth/Role.cs
@@ -1,8 +1,0 @@
-using System;
-using Microsoft.AspNetCore.Identity;
-
-namespace LinuxLudo.API.Domain.Models.Auth
-{
-    public class Role : IdentityRole<Guid>
-    { }
-}
