+++ src/LinuxLudo.API/Domain/Repositories/IGameResultRepository.cs
@@ -3,7 +3,7 @@ using LinuxLudo.API.Domain.Models;
 
 namespace LinuxLudo.API.Domain.Repositories
 {
-    public interface IPlayerStatsRepository : IRepository<PlayerStats>
+    public interface IGameResultRepository : IRepository<GameResult>
     {
 
     }
