+++ src/LinuxLudo.API/Domain/Repositories/IUnitOfWork.cs
@@ -6,7 +6,7 @@ namespace LinuxLudo.API.Domain.Repositories
     public interface IUnitOfWork : IDisposable
     {
         IGameRepository Games { get; }
-        IPlayerStatsRepository PlayerStats { get; }
+        IGameResultRepository GameResult { get; }
         Task<int> CommitAsync();
     }
 }
