--- src/LinuxLudo.API/Domain/Resources/Auth/SignInResource.cs
@@ -1,12 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace LinuxLudo.API.Domain.Resources.Auth
-{
-    public class SignInResource
-    {
-        [Required]
-        public string Username { get; set; }
-        [Required]
-        public string Password { get; set; }
-    }
-}
