--- src/LinuxLudo.API/Domain/Resources/Auth/SignUpResource.cs
@@ -1,12 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace LinuxLudo.API.Domain.Resources.Auth
-{
-    public class SignUpResource
-    {
-        [Required]
-        public string Username { get; set; }
-        [Required]
-        public string Password { get; set; }
-    }
-}
