--- src/LinuxLudo.API/Domain/Resources/Response/TokenResponse.cs
@@ -1,12 +1,0 @@
-namespace LinuxLudo.API.Domain.Resources.Response
-{
-    public class TokenResponse
-    {
-        public string Token { get; }
-
-        public TokenResponse(string token)
-        {
-            Token = token;
-        }
-    }
-}
