--- src/LinuxLudo.API/Extensions/AuthExtentions.cs
@@ -1,43 +1,0 @@
-using System.Text;
-using LinuxLudo.API.Settings;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.IdentityModel.Tokens;
-
-namespace LinuxLudo.API.Extensions
-{
-    public static class AuthExtensions
-    {
-        public static IServiceCollection AddAuth(this IServiceCollection services, JwtSettings jwtSettings)
-        {
-            services
-                .AddAuthorization()
-                .AddAuthentication(opts =>
-                {
-                    opts.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                    opts.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
-                    opts.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(opts =>
-                {
-                    opts.RequireHttpsMetadata = false;
-                    opts.TokenValidationParameters = new TokenValidationParameters()
-                    {
-                        ValidIssuer = jwtSettings.Issuer,
-                        ValidAudience = jwtSettings.Issuer,
-                        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtSettings.Secret))
-                    };
-                });
-
-            return services;
-        }
-
-        public static IApplicationBuilder UseAuth(this IApplicationBuilder app)
-        {
-            app.UseAuthentication();
-            app.UseAuthorization();
-            return app;
-        }
-    }
-}
