--- src/LinuxLudo.API/Settings/JwtSettings.cs
@@ -1,9 +1,0 @@
-namespace LinuxLudo.API.Settings
-{
-    public class JwtSettings
-    {
-        public string Issuer { get; set; }
-        public string Secret { get; set; }
-        public int ExpirationInDays { get; set; }
-    }
-}
