+++ src/LinuxLudo.API/Startup.cs
@@ -1,22 +1,16 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json.Serialization;
 using LinuxLudo.API.Database.Context;
 using LinuxLudo.API.Database.Repositories;
-using LinuxLudo.API.Domain.Models.Auth;
 using LinuxLudo.API.Domain.Repositories;
 using LinuxLudo.API.Domain.Response;
 using LinuxLudo.API.Domain.Services;
 using LinuxLudo.API.Extensions;
 using LinuxLudo.API.Hubs;
-using LinuxLudo.API.Middleware;
 using LinuxLudo.API.Services;
-using LinuxLudo.API.Settings;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.ResponseCompression;
 using Microsoft.EntityFrameworkCore;
@@ -39,16 +33,7 @@ namespace LinuxLudo.API
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.Configure<JwtSettings>(Configuration.GetSection("Jwt"));
-            var jwtSettings = Configuration.GetSection("Jwt").Get<JwtSettings>();
             services.AddDbContext<AppDbContext>(opts => opts.UseNpgsql(Configuration.GetConnectionString("Default")));
-            services.AddIdentity<User, Role>(opts =>
-                {
-                    opts.Password.RequireNonAlphanumeric = false;
-                    opts.Password.RequireUppercase = false;
-                })
-                .AddEntityFrameworkStores<AppDbContext>()
-                .AddDefaultTokenProviders();
             services.AddAutoMapper(typeof(Startup));
             services.AddSignalR(opts =>
             {
@@ -74,40 +59,11 @@ namespace LinuxLudo.API
             {
                 c.SwaggerDoc("v1", new OpenApiInfo { Title = "LinuxLudo.API", Version = "v1" });
 
-                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme()
-                {
-                    Description = "JWT containing userid claim",
-                    Name = "Authorization",
-                    In = ParameterLocation.Header,
-                    Type = SecuritySchemeType.ApiKey,
-                });
-
-                var security = new OpenApiSecurityRequirement()
-                {
-                    {
-                    new OpenApiSecurityScheme()
-                    {
-                        Reference = new OpenApiReference()
-                        {
-                            Id = "Bearer",
-                            Type = ReferenceType.SecurityScheme
-                        },
-                        UnresolvedReference = true
-                    },
-                    new List<string>()
-                    }
-                };
-
-                c.AddSecurityRequirement(security);
             });
 
             services.AddScoped<IUnitOfWork, UnitOfWork>();
-            services.AddTransient<IJwtService, JwtService>();
-            services.AddTransient<IAuthService, AuthService>();
             services.AddTransient<IGameService, GameService>();
             services.AddSingleton<IGameHubRepository, GameHubRepository>();
-
-            services.AddAuth(jwtSettings);
 
             services.AddCors(options =>
             {
@@ -140,7 +96,6 @@ namespace LinuxLudo.API
             app.UseCors("Open");
             app.ApplyCustomMiddleware();
 
-            app.UseAuth();
             app.UseEndpoints(endpoints =>
             {
                 endpoints.MapControllers();
