+++ src/LinuxLudo.Core/Models/OpenGame.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace LinuxLudo.Web.Domain.Models
+namespace LinuxLudo.Core.Models
 {
     public class OpenGame
     {
